Update AnimationSetup plugin and fix references.
Remake FirstPerson portrayal in native code instead of BP while I'm at reparenting it with the new class name. Idk even when the class is this simple to the point that it is empty, it still feels easier just having it in text in c plus plus lol.

Update GameCore plugin.